Created a CompletedChoreRepository and refactored ChoreCompletionRequest to be ChoreCompletionDTO
